requirements.txt
# PPTX Automation Tool - Requirements

# Core dependencies
python-pptx>=0.6.21
jinja2>=3.1.2

# Optional: For audio transcription
# openai-whisper>=20231117
# faster-whisper>=0.9.0

# Optional: For LLM-based content generation
# openai>=1.3.0
# anthropic>=0.7.0

# Optional: For text processing
# nltk>=3.8.1
# spacy>=3.7.0

# Optional: For audio processing
# pydub>=0.25.1
# soundfile>=0.12.1

# Development dependencies
# pytest>=7.4.3
# black>=23.12.0
# flake8>=6.1.0
